Fix page turn position/scaling bugs
PageView pre-builds adjacent pages, so when a bunch of pages using
the same controller get created, the controller's internal state will be
overwritten with the state of the latest page child created. This,
combined with the fact that PhotoView's 'contained' scale recalculates
the child's bounds asynchronously after the network image is loaded and
the size is known, means often times the controller's initial scale and
bounds are wrong if the most recent page preloaded by PageView contains
a different size/aspect ratio than the current page.

Hopefully this fixes the remaining issues with errant initial sizes and
image positions.