Replace list title view with paginated view
Infinite scroll view doesn't work well when the list is modifiable, and
when preferably having the list react when modified in order to save a
user refresh or extra API calls.

Making the list react to a successful update without requerying the list
involves making CustomList a pseudo-mutable using ChangeNotifier, and
signaling when the internal list changes, despite the identity of the
neither the CustomList nor the internal list changing, then updating the
state of the widget's view of the list.