Open on site deeplink workaround
Takes advantage of the fact that the md/cb servers still redirect http
to https to workaround the fact that url_launcher doesn't work/in-app
browser mode isn't respected or is ignored/preempted when deeplinks are
enabled on those domains, and gagaku only registers https as the scheme.

Not a good long term solution and needs to be amended when there is a
proper fix.